Fix dependency loop between poll_appengine and server
