+++ config/__init__.py
@@ -1,0 +1,2 @@
+def valve_position(valve_no):
+  return -9.625 - 1.875 * valve_no
